Moved bunch of tasks outside of avionics.cpp and discovered a bunch of bugs
