+++ avionics/common/CMakeLists.txt
@@ -7,6 +7,7 @@ set(common_files
     ${CMAKE_CURRENT_LIST_DIR}/src/sensor_collection.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/radio.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/radio_queue.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/src/rocket.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/status.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/XBee.cpp
 
@@ -18,7 +19,12 @@ set(common_files
     ${CMAKE_CURRENT_LIST_DIR}/src/states/pressure_delay.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/states/standby.cpp
 
+    ${CMAKE_CURRENT_LIST_DIR}/src/tasks/main_tasks.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/src/tasks/led_blinker.cpp
+
     ${CMAKE_CURRENT_LIST_DIR}/src/hardware/ignitor.cpp
+
+    ${CMAKE_CURRENT_LIST_DIR}/lib/CalcHelpers/barometric.cpp
 )
 target_sources(
     TantalusStage1
