+++ avionics/common/lib/CalcHelpers/barometric.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <cmath>
+
+float pressureToAltitude(float const pres);
