+++ avionics/common/src/rocket.cpp
@@ -1,0 +1,2 @@
+#include "rocket.h"
+constexpr char Rocket::LOG_FILE_NAME[];
