+++ avionics/common/src/tasks/led_blinker.cpp
@@ -1,0 +1,3 @@
+#include "tasks/led_blinker.hpp"
+bool LEDBlinker::led_on = true;
+constexpr Hal::ms LEDBlinker::freq;
