+++ avionics/common/src/tasks/main_tasks.cpp
@@ -1,0 +1,2 @@
+#include "tasks/main_tasks.hpp"
+constexpr Hal::ms RadioTxBulk::freq;
