+++ avionics/rockets/tantalus_stage_2/include/calculations.h
@@ -5,7 +5,7 @@
 
 #include <cmath> //std::abs
 
-#include "CalcHelpers/barometric.h"
+#include "CalcHelpers/barometric.hpp"
 #include "CalcHelpers/exponential_moving_avg.h"
 #include "calculations_interface.h"
 #include "sensor_collection.h"
